+++ tools/awps-tunnel/client/src/components/Connector.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { render } from '@testing-library/react';
 import { Connector } from './Connector';
 import { ConnectionStatus } from '../models';