+++ tools/awps-tunnel/server/CHANGELOG.md
@@ -1,9 +1,19 @@
 # Changelog
 
+## [1.0.0-beta.10] - 2024-09-21
+### Improved
+- Some improvements to the REST API tab
+
+## [1.0.0-beta.9] - 2024-09-20
+### Improved
+- Add a format view of the upstream requests and response
+- Add a InvokeService tab in the app server pannel
+- Support AzurePowerShellCredential
+
 ## [1.0.0-beta.8] - 2024-04-10
 ### Fixed
 - Fix the issue when connection string still set, the tool still tries to read endpoint setting
-- 
+
 ## [1.0.0-beta.7] - 2024-04-09
 ### Added
 - Support passing --connection option when running the tunnel